node: MockApiClient available during development
