+++ node/src/cli.rs
@@ -1,10 +1,7 @@
-use std::sync::Arc;
-
 use anyhow::Context;
 use argh::FromArgs;
 use common::rng::SysRng;
 
-use crate::api::LexeApiClient;
 use crate::bitcoind_client::BitcoindRpcInfo;
 use crate::init;
 use crate::provision::{provision, LexeRunner};
@@ -82,6 +79,10 @@ pub struct StartCommand {
     /// protocol://host:port of the runner.
     #[argh(option, default = "DEFAULT_RUNNER_URL.into()")]
     pub runner_url: String,
+
+    /// whether to use a mock API client. Only available during development.
+    #[argh(switch, short = 'm')]
+    pub mock: bool,
 }
 
 /// Provision a new Lexe node for a user
@@ -125,11 +126,7 @@ impl Args {
                     .build()
                     .expect("Failed to build tokio runtime");
                 let mut rng = SysRng::new();
-                let api = Arc::new(LexeApiClient::new(
-                    args.backend_url.clone(),
-                    args.runner_url.clone(),
-                ));
-                rt.block_on(init::start_ldk(&mut rng, args, api))
+                rt.block_on(init::start_ldk(&mut rng, args))
                     .context("Error running node")
             }
             Command::Provision(args) => {
